backend/requirements.txt
fastapi
numpy
opencv-python
pillow
python-multipart
uvicorn
streamlit

